refactor: 콘솔 관련 IO 로직 view/console 폴더로 이동
