+++ src/js/view/console/Output.js
